@@ -16,8 +16,8 @@
     "format:all": "yarn format \"**/*.{js,ts,tsx,json}\"",
     "test": "jest",
     "type-check": "tsc --noEmit",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "lint-staged": {
     "*.{js,ts,tsx,json}": [
@@ -29,7 +29,7 @@
   },
   "dependencies": {
     "@apollo/client": "^3.4.10",
-    "@headlessui/react": "^1.4.0",
+    "@headlessui/react": "0.0.0-insiders.d55c77e",
     "@primer/octicons-react": "^15.0.0",
     "@swc/core": "^1.3.24",
     "@vanilla-extract/css": "^1.3.0",
@@ -37,28 +37,29 @@
     "cross-fetch": "^3.1.4",
     "destyle.css": "^2.0.2",
     "jotai": "^1.1.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0"
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.9.0"
   },
   "devDependencies": {
     "@babel/core": "7.21.0",
-    "@storybook/addon-actions": "6.5.16",
-    "@storybook/addon-essentials": "6.5.16",
-    "@storybook/addon-interactions": "6.5.16",
-    "@storybook/addon-links": "6.5.16",
-    "@storybook/builder-vite": "0.4.2",
-    "@storybook/jest": "0.0.10",
-    "@storybook/react": "6.5.16",
-    "@storybook/testing-library": "0.0.13",
-    "@storybook/testing-react": "1.3.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/react-hooks": "^7.0.1",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.8",
+    "@storybook/addon-actions": "7.0.0-rc.5",
+    "@storybook/addon-essentials": "7.0.0-rc.5",
+    "@storybook/addon-interactions": "7.0.0-rc.5",
+    "@storybook/addon-links": "7.0.0-rc.5",
+    "@storybook/addon-mdx-gfm": "7.0.0-rc.5",
+    "@storybook/jest": "0.0.11-next.0",
+    "@storybook/react": "7.0.0-rc.5",
+    "@storybook/react-vite": "7.0.0-rc.5",
+    "@storybook/testing-library": "0.0.14-next.1",
+    "@storybook/testing-react": "2.0.0-next.0",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/react-hooks": "8.0.1",
+    "@testing-library/user-event": "14.4.3",
+    "@types/react": "18.0.28",
+    "@types/react-dom": "18.0.11",
+    "@types/react-router-dom": "5.3.3",
     "@vanilla-extract/vite-plugin": "^2.0.0",
     "@vitejs/plugin-react-refresh": "^1.3.1",
     "@vitejs/plugin-react-swc": "3.2.0",
@@ -66,6 +67,7 @@
     "graphql": "^15.5.1",
     "identity-obj-proxy": "^3.0.0",
     "react-test-renderer": "^17.0.2",
+    "storybook": "7.0.0-rc.5",
     "vite": "4.1.3"
   }
 }
