@@ -1,25 +1,17 @@
-import React, { FC, ReactNode, Suspense } from 'react';
-import { Route, RouteProps, Redirect } from 'react-router-dom';
+import { FC, PropsWithChildren, Suspense } from 'react';
+import { Route, RouteProps, Navigate } from 'react-router-dom';
 import { useAtom } from 'jotai';
 import { isLoggedInAtom } from '../../../jotai';
 import { AuthProvider } from '../../AuthProvider';
 
-type PrivateRouteProps = RouteProps & {
-  children?: ReactNode;
-};
-
-export const PrivateRoute: FC<PrivateRouteProps> = ({ children, ...rest }) => {
+export const PrivateRoute: FC<PropsWithChildren> = ({ children }) => {
   const [isLoggedIn] = useAtom(isLoggedInAtom);
 
-  return (
-    <Route {...rest}>
-      {isLoggedIn ? (
-        <Suspense fallback="Loading...">
-          <AuthProvider>{children}</AuthProvider>
-        </Suspense>
-      ) : (
-        <Redirect to="/login" />
-      )}
-    </Route>
+  return isLoggedIn ? (
+    <Suspense fallback="Loading...">
+      <AuthProvider>{children}</AuthProvider>
+    </Suspense>
+  ) : (
+    <Navigate to="/login" replace />
   );
 };
