@@ -1,5 +1,5 @@
 import React, { useCallback } from 'react';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { Button } from '../../components';
 import { useSettings } from '../../hooks';
 import { BaseLayout } from '../../layouts/BaseLayout';
@@ -7,16 +7,16 @@ import { completeButtonStyle, containerStyle, titleStyle } from './styles.css';
 import { SelectRepositoryContainer } from '../../containers/SelectRepositoryContainer/SelectRepositoryContainer';
 
 export const SelectRepositoryPage = () => {
-  const history = useHistory();
+  const navigate = useNavigate();
   const { settings } = useSettings();
 
   const handleClick = useCallback(() => {
     if (settings.subscribedRepositories.length === 0) {
       alert('レビュー対象のリポジトリを1つ以上選択してください。');
     } else {
-      history.replace('/');
+      navigate('/', { replace: true });
     }
-  }, [settings, history]);
+  }, [settings, navigate]);
 
   return (
     <BaseLayout>
