@@ -1,15 +1,15 @@
 const { mergeConfig } = require('vite');
-
 module.exports = {
-  stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
+  stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-interactions',
+    '@storybook/addon-mdx-gfm',
   ],
-  framework: '@storybook/react',
-  core: {
-    builder: '@storybook/builder-vite',
+  framework: {
+    name: '@storybook/react-vite',
+    options: {},
   },
   features: {
     storyStoreV7: true,
@@ -19,5 +19,8 @@ module.exports = {
     return mergeConfig(config, {
       plugins: [require('@vanilla-extract/vite-plugin').vanillaExtractPlugin()],
     });
+  },
+  docs: {
+    autodocs: true,
   },
 };
